Get fetchAllPartitionsWorking sorta
Wowzors, so the connector code works as it should. But simply changing the request parameters and resending is not even possible within the custom connector framework. So there is no way to do subsequent requests, unless we were to recompose a new one based on the old one.

TLDR, using multiple api calls within a single connector call may not be a viable approach without even more work.