In a working state as a custom connector now.
Woot. Was able to get this working and tested as a custom connector.

I did have to nuke the other ops b/c the swaggerjson was fraught with issues. We will need to do a careful merge later with Joe's changes.

Also, the custom connector upload did not like the recursive schema of the "Assertion" parameter of AssertAll, so I dumbed it down to a plain object.