Tchee/ppunit assertions (#12)
* AssertAll progress

- Progress towards an "AssertAll" method for doing all assertions in one go and return an array of messages so the flows don't have to be so messy.

* Update script.csx

Finish ability to assert type, null and empty. I think this needs reworked one more time to allow chaining of AND's or OR's, otherwise we have to assume AND, which would force the user to do multiple AssertAll statements, which could also be a viable options depending on how difficult it is to add support for chaning assertions with both AND and OR.

* Update script.csx

Add type detection for float.

* Update script.csx

Attempt to get comparison operators working. Still need to add support for grouping assertions with an AND or an OR operator for more powerful assertions without the user having to to make multiple calls to AssertAll and then string them all together in powerapps.

* Update script.csx

- Comments for how to build out logical operator tomorrow hopefully.

* Update script.csx

This works from a customized debug console. Still waiting to see if this work as a custom connector.

* Update script.csx

- Tweaks to AssertAll to get more accurate results. Clear errors if Passed equals true.

* In a working state as a custom connector now.

Woot. Was able to get this working and tested as a custom connector.

I did have to nuke the other ops b/c the swaggerjson was fraught with issues. We will need to do a careful merge later with Joe's changes.

Also, the custom connector upload did not like the recursive schema of the "Assertion" parameter of AssertAll, so I dumbed it down to a plain object.

* Update script.csx

- fix excessive short circuiting.
- fix actual response structure not matching the api description.

* Remove payload object.

Remove payload and directly embed values in the Left/RightExpression properties. Using JToken seems to have been the key to cracking this without incurring a massive rewrite.

* Update script.csx

Code cleanup.

* missing codefixes + documentation.

- I guess one of the code fixes never got committed.
- Update documentation.